Change access modifier of commands, split nu-plugin, move stuff out of nu-test-supp
