+++ crates/nu-command/src/commands.rs
@@ -1,295 +1,294 @@
 #[macro_use]
-pub(crate) mod macros;
+pub mod macros;
 
 mod from_delimited_data;
 mod to_delimited_data;
 
-pub(crate) mod ansi;
-pub(crate) mod append;
-pub(crate) mod args;
+pub mod ansi;
+pub mod append;
+pub mod args;
 pub mod autoenv;
-pub(crate) mod autoenv_trust;
-pub(crate) mod autoenv_untrust;
-pub(crate) mod autoview;
-pub(crate) mod benchmark;
-pub(crate) mod build_string;
-pub(crate) mod cal;
-pub(crate) mod cd;
-pub(crate) mod char_;
-pub(crate) mod chart;
-pub(crate) mod classified;
+pub mod autoenv_trust;
+pub mod autoenv_untrust;
+pub mod autoview;
+pub mod benchmark;
+pub mod build_string;
+pub mod cal;
+pub mod cd;
+pub mod char_;
+pub mod chart;
+pub mod classified;
 #[cfg(feature = "clipboard-cli")]
-pub(crate) mod clip;
-pub(crate) mod compact;
-pub(crate) mod config;
-pub(crate) mod constants;
-pub(crate) mod cp;
-pub(crate) mod date;
-pub(crate) mod debug;
-pub(crate) mod def;
-pub(crate) mod default;
+pub mod clip;
+pub mod compact;
+pub mod config;
+pub mod constants;
+pub mod cp;
+pub mod date;
+pub mod debug;
+pub mod def;
+pub mod default;
 pub mod default_context;
-pub(crate) mod describe;
-pub(crate) mod do_;
-pub(crate) mod drop;
-pub(crate) mod du;
-pub(crate) mod each;
-pub(crate) mod echo;
-pub(crate) mod empty;
-pub(crate) mod enter;
-pub(crate) mod every;
-pub(crate) mod exec;
-pub(crate) mod exit;
-pub(crate) mod first;
-pub(crate) mod flatten;
-pub(crate) mod format;
-pub(crate) mod from;
-pub(crate) mod from_csv;
-pub(crate) mod from_eml;
-pub(crate) mod from_ics;
-pub(crate) mod from_ini;
-pub(crate) mod from_json;
-pub(crate) mod from_ods;
-pub(crate) mod from_ssv;
-pub(crate) mod from_toml;
-pub(crate) mod from_tsv;
-pub(crate) mod from_url;
-pub(crate) mod from_vcf;
-pub(crate) mod from_xlsx;
-pub(crate) mod from_xml;
-pub(crate) mod from_yaml;
-pub(crate) mod get;
-pub(crate) mod group_by;
-pub(crate) mod group_by_date;
-pub(crate) mod hash_;
-pub(crate) mod headers;
-pub(crate) mod help;
-pub(crate) mod histogram;
-pub(crate) mod history;
-pub(crate) mod if_;
-pub(crate) mod insert;
-pub(crate) mod into_int;
-pub(crate) mod keep;
-pub(crate) mod last;
-pub(crate) mod length;
-pub(crate) mod let_;
-pub(crate) mod let_env;
-pub(crate) mod lines;
-pub(crate) mod ls;
-pub(crate) mod math;
-pub(crate) mod merge;
-pub(crate) mod mkdir;
-pub(crate) mod move_;
-pub(crate) mod next;
-pub(crate) mod nth;
-pub(crate) mod nu;
-pub(crate) mod open;
-pub(crate) mod parse;
-pub(crate) mod path;
-pub(crate) mod pivot;
-pub(crate) mod prepend;
-pub(crate) mod prev;
-pub(crate) mod pwd;
-pub(crate) mod random;
-pub(crate) mod range;
-pub(crate) mod reduce;
-pub(crate) mod reject;
-pub(crate) mod rename;
-pub(crate) mod reverse;
-pub(crate) mod rm;
-pub(crate) mod roll;
-pub(crate) mod rotate;
-pub(crate) mod run_external;
-pub(crate) mod save;
-pub(crate) mod select;
-pub(crate) mod seq;
-pub(crate) mod seq_dates;
-pub(crate) mod shells;
-pub(crate) mod shuffle;
-pub(crate) mod size;
-pub(crate) mod skip;
-pub(crate) mod sleep;
-pub(crate) mod sort_by;
-pub(crate) mod source;
-pub(crate) mod split;
-pub(crate) mod split_by;
-pub(crate) mod str_;
-pub(crate) mod table;
-pub(crate) mod tags;
-pub(crate) mod termsize;
-pub(crate) mod to;
-pub(crate) mod to_csv;
-pub(crate) mod to_html;
-pub(crate) mod to_json;
-pub(crate) mod to_md;
-pub(crate) mod to_toml;
-pub(crate) mod to_tsv;
-pub(crate) mod to_url;
-pub(crate) mod to_xml;
-pub(crate) mod to_yaml;
-pub(crate) mod uniq;
-pub(crate) mod update;
-pub(crate) mod url_;
-pub(crate) mod version;
-pub(crate) mod where_;
-pub(crate) mod which_;
-pub(crate) mod with_env;
-pub(crate) mod wrap;
+pub mod describe;
+pub mod do_;
+pub mod drop;
+pub mod du;
+pub mod each;
+pub mod echo;
+pub mod empty;
+pub mod enter;
+pub mod every;
+pub mod exec;
+pub mod exit;
+pub mod first;
+pub mod flatten;
+pub mod format;
+pub mod from;
+pub mod from_csv;
+pub mod from_eml;
+pub mod from_ics;
+pub mod from_ini;
+pub mod from_json;
+pub mod from_ods;
+pub mod from_ssv;
+pub mod from_toml;
+pub mod from_tsv;
+pub mod from_url;
+pub mod from_vcf;
+pub mod from_xlsx;
+pub mod from_xml;
+pub mod from_yaml;
+pub mod get;
+pub mod group_by;
+pub mod group_by_date;
+pub mod hash_;
+pub mod headers;
+pub mod help;
+pub mod histogram;
+pub mod history;
+pub mod if_;
+pub mod insert;
+pub mod into_int;
+pub mod keep;
+pub mod last;
+pub mod length;
+pub mod let_;
+pub mod let_env;
+pub mod lines;
+pub mod ls;
+pub mod math;
+pub mod merge;
+pub mod mkdir;
+pub mod move_;
+pub mod next;
+pub mod nth;
+pub mod nu;
+pub mod open;
+pub mod parse;
+pub mod path;
+pub mod pivot;
+pub mod prepend;
+pub mod prev;
+pub mod pwd;
+pub mod random;
+pub mod range;
+pub mod reduce;
+pub mod reject;
+pub mod rename;
+pub mod reverse;
+pub mod rm;
+pub mod roll;
+pub mod rotate;
+pub mod run_external;
+pub mod save;
+pub mod select;
+pub mod seq;
+pub mod seq_dates;
+pub mod shells;
+pub mod shuffle;
+pub mod size;
+pub mod skip;
+pub mod sleep;
+pub mod sort_by;
+pub mod source;
+pub mod split;
+pub mod split_by;
+pub mod str_;
+pub mod table;
+pub mod tags;
+pub mod termsize;
+pub mod to;
+pub mod to_csv;
+pub mod to_html;
+pub mod to_json;
+pub mod to_md;
+pub mod to_toml;
+pub mod to_tsv;
+pub mod to_url;
+pub mod to_xml;
+pub mod to_yaml;
+pub mod uniq;
+pub mod update;
+pub mod url_;
+pub mod version;
+pub mod where_;
+pub mod which_;
+pub mod with_env;
+pub mod wrap;
 
-pub(crate) use autoview::Autoview;
-pub(crate) use cd::Cd;
+pub use autoview::Autoview;
+pub use cd::Cd;
 
-pub(crate) use ansi::Ansi;
-pub(crate) use ansi::AnsiStrip;
-pub(crate) use append::Command as Append;
-pub(crate) use autoenv::Autoenv;
-pub(crate) use autoenv_trust::AutoenvTrust;
-pub(crate) use autoenv_untrust::AutoenvUnTrust;
-pub(crate) use benchmark::Benchmark;
-pub(crate) use build_string::BuildString;
-pub(crate) use cal::Cal;
-pub(crate) use char_::Char;
-pub(crate) use chart::Chart;
-pub(crate) use compact::Compact;
-pub(crate) use config::{
+pub use ansi::Ansi;
+pub use ansi::AnsiStrip;
+pub use append::Command as Append;
+pub use autoenv::Autoenv;
+pub use autoenv_trust::AutoenvTrust;
+pub use autoenv_untrust::AutoenvUnTrust;
+pub use benchmark::Benchmark;
+pub use build_string::BuildString;
+pub use cal::Cal;
+pub use char_::Char;
+pub use chart::Chart;
+pub use compact::Compact;
+pub use config::{
     Config, ConfigClear, ConfigGet, ConfigPath, ConfigRemove, ConfigSet, ConfigSetInto,
 };
-pub(crate) use cp::Cpy;
-pub(crate) use date::{Date, DateFormat, DateListTimeZone, DateNow, DateToTable, DateToTimeZone};
-pub(crate) use debug::Debug;
-pub(crate) use def::Def;
-pub(crate) use default::Default;
-pub(crate) use describe::Describe;
-pub(crate) use do_::Do;
-pub(crate) use drop::{Drop, DropColumn};
-pub(crate) use du::Du;
-pub(crate) use each::Each;
-pub(crate) use each::EachGroup;
-pub(crate) use each::EachWindow;
-pub(crate) use echo::Echo;
-pub(crate) use empty::Command as Empty;
-pub(crate) use if_::If;
-pub(crate) use nu::NuPlugin;
-pub(crate) use update::Command as Update;
-pub(crate) mod kill;
-pub(crate) use kill::Kill;
-pub(crate) mod clear;
-pub(crate) use clear::Clear;
-pub(crate) mod touch;
-pub(crate) use enter::Enter;
-pub(crate) use every::Every;
-pub(crate) use exec::Exec;
-pub(crate) use exit::Exit;
-pub(crate) use first::First;
-pub(crate) use flatten::Command as Flatten;
-pub(crate) use format::{FileSize, Format};
-pub(crate) use from::From;
-pub(crate) use from_csv::FromCsv;
-pub(crate) use from_eml::FromEml;
-pub(crate) use from_ics::FromIcs;
-pub(crate) use from_ini::FromIni;
-pub(crate) use from_json::FromJson;
-pub(crate) use from_ods::FromOds;
-pub(crate) use from_ssv::FromSsv;
-pub(crate) use from_toml::FromToml;
-pub(crate) use from_tsv::FromTsv;
-pub(crate) use from_url::FromUrl;
-pub(crate) use from_vcf::FromVcf;
-pub(crate) use from_xlsx::FromXlsx;
-pub(crate) use from_xml::FromXml;
-pub(crate) use from_yaml::FromYaml;
-pub(crate) use from_yaml::FromYml;
-pub(crate) use get::Command as Get;
-pub(crate) use group_by::Command as GroupBy;
-pub(crate) use group_by_date::GroupByDate;
-pub(crate) use hash_::{Hash, HashBase64, HashMd5};
-pub(crate) use headers::Headers;
-pub(crate) use help::Help;
-pub(crate) use histogram::Histogram;
-pub(crate) use history::History;
-pub(crate) use insert::Command as Insert;
-pub(crate) use into_int::IntoInt;
-pub(crate) use keep::{Keep, KeepUntil, KeepWhile};
-pub(crate) use last::Last;
-pub(crate) use length::Length;
-pub(crate) use let_::Let;
-pub(crate) use let_env::LetEnv;
-pub(crate) use lines::Lines;
-pub(crate) use ls::Ls;
-pub(crate) use math::{
+pub use cp::Cpy;
+pub use date::{Date, DateFormat, DateListTimeZone, DateNow, DateToTable, DateToTimeZone};
+pub use debug::Debug;
+pub use def::Def;
+pub use default::Default;
+pub use describe::Describe;
+pub use do_::Do;
+pub use drop::{Drop, DropColumn};
+pub use du::Du;
+pub use each::Each;
+pub use each::EachGroup;
+pub use each::EachWindow;
+pub use echo::Echo;
+pub use empty::Command as Empty;
+pub use if_::If;
+pub use nu::NuPlugin;
+pub use update::Command as Update;
+pub mod kill;
+pub use kill::Kill;
+pub mod clear;
+pub use clear::Clear;
+pub mod touch;
+pub use enter::Enter;
+pub use every::Every;
+pub use exec::Exec;
+pub use exit::Exit;
+pub use first::First;
+pub use flatten::Command as Flatten;
+pub use format::{FileSize, Format};
+pub use from::From;
+pub use from_csv::FromCsv;
+pub use from_eml::FromEml;
+pub use from_ics::FromIcs;
+pub use from_ini::FromIni;
+pub use from_json::FromJson;
+pub use from_ods::FromOds;
+pub use from_ssv::FromSsv;
+pub use from_toml::FromToml;
+pub use from_tsv::FromTsv;
+pub use from_url::FromUrl;
+pub use from_vcf::FromVcf;
+pub use from_xlsx::FromXlsx;
+pub use from_xml::FromXml;
+pub use from_yaml::FromYaml;
+pub use from_yaml::FromYml;
+pub use get::Command as Get;
+pub use group_by::Command as GroupBy;
+pub use group_by_date::GroupByDate;
+pub use hash_::{Hash, HashBase64, HashMd5};
+pub use headers::Headers;
+pub use help::Help;
+pub use histogram::Histogram;
+pub use history::History;
+pub use insert::Command as Insert;
+pub use into_int::IntoInt;
+pub use keep::{Keep, KeepUntil, KeepWhile};
+pub use last::Last;
+pub use length::Length;
+pub use let_::Let;
+pub use let_env::LetEnv;
+pub use lines::Lines;
+pub use ls::Ls;
+
+pub use math::{
     Math, MathAbs, MathAverage, MathCeil, MathEval, MathFloor, MathMaximum, MathMedian,
     MathMinimum, MathMode, MathProduct, MathRound, MathSqrt, MathStddev, MathSummation,
     MathVariance,
 };
-pub(crate) use merge::Merge;
-pub(crate) use mkdir::Mkdir;
-pub(crate) use move_::{Move, Mv};
-pub(crate) use next::Next;
-pub(crate) use nth::Nth;
-pub(crate) use open::Open;
-pub(crate) use parse::Parse;
-pub(crate) use path::{
+pub use merge::Merge;
+pub use mkdir::Mkdir;
+pub use move_::{Move, Mv};
+pub use next::Next;
+pub use nth::Nth;
+pub use open::Open;
+pub use parse::Parse;
+pub use path::{
     PathBasename, PathCommand, PathDirname, PathExists, PathExpand, PathExtension, PathFilestem,
     PathJoin, PathType,
 };
-pub(crate) use pivot::Pivot;
-pub(crate) use prepend::Prepend;
-pub(crate) use prev::Previous;
-pub(crate) use pwd::Pwd;
+pub use pivot::Pivot;
+pub use prepend::Prepend;
+pub use prev::Previous;
+pub use pwd::Pwd;
 #[cfg(feature = "uuid_crate")]
-pub(crate) use random::RandomUUID;
-pub(crate) use random::{
-    Random, RandomBool, RandomChars, RandomDecimal, RandomDice, RandomInteger,
-};
-pub(crate) use range::Range;
-pub(crate) use reduce::Reduce;
-pub(crate) use reject::Reject;
-pub(crate) use rename::Rename;
-pub(crate) use reverse::Reverse;
-pub(crate) use rm::Remove;
-pub(crate) use roll::{Roll, RollColumn, RollUp};
-pub(crate) use rotate::{Rotate, RotateCounterClockwise};
-pub(crate) use run_external::RunExternalCommand;
-pub(crate) use save::Save;
-pub(crate) use select::Command as Select;
-pub(crate) use seq::Seq;
-pub(crate) use seq_dates::SeqDates;
-pub(crate) use shells::Shells;
-pub(crate) use shuffle::Shuffle;
-pub(crate) use size::Size;
-pub(crate) use skip::{Skip, SkipUntil, SkipWhile};
-pub(crate) use sleep::Sleep;
-pub(crate) use sort_by::SortBy;
-pub(crate) use source::Source;
-pub(crate) use split::{Split, SplitChars, SplitColumn, SplitRow};
-pub(crate) use split_by::SplitBy;
-pub(crate) use str_::{
+pub use random::RandomUUID;
+pub use random::{Random, RandomBool, RandomChars, RandomDecimal, RandomDice, RandomInteger};
+pub use range::Range;
+pub use reduce::Reduce;
+pub use reject::Reject;
+pub use rename::Rename;
+pub use reverse::Reverse;
+pub use rm::Remove;
+pub use roll::{Roll, RollColumn, RollUp};
+pub use rotate::{Rotate, RotateCounterClockwise};
+pub use run_external::RunExternalCommand;
+pub use save::Save;
+pub use select::Command as Select;
+pub use seq::Seq;
+pub use seq_dates::SeqDates;
+pub use shells::Shells;
+pub use shuffle::Shuffle;
+pub use size::Size;
+pub use skip::{Skip, SkipUntil, SkipWhile};
+pub use sleep::Sleep;
+pub use sort_by::SortBy;
+pub use source::Source;
+pub use split::{Split, SplitChars, SplitColumn, SplitRow};
+pub use split_by::SplitBy;
+pub use str_::{
     Str, StrCamelCase, StrCapitalize, StrCollect, StrContains, StrDowncase, StrEndsWith,
     StrFindReplace, StrFrom, StrIndexOf, StrKebabCase, StrLPad, StrLength, StrPascalCase, StrRPad,
     StrReverse, StrScreamingSnakeCase, StrSnakeCase, StrStartsWith, StrSubstring, StrToDatetime,
     StrToDecimal, StrToInteger, StrTrim, StrTrimLeft, StrTrimRight, StrUpcase,
 };
-pub(crate) use table::Table;
-pub(crate) use tags::Tags;
-pub(crate) use termsize::TermSize;
-pub(crate) use to::To;
-pub(crate) use to_csv::ToCsv;
-pub(crate) use to_html::ToHtml;
-pub(crate) use to_json::ToJson;
-pub(crate) use to_md::Command as ToMarkdown;
-pub(crate) use to_toml::ToToml;
-pub(crate) use to_tsv::ToTsv;
-pub(crate) use to_url::ToUrl;
-pub(crate) use to_xml::ToXml;
-pub(crate) use to_yaml::ToYaml;
-pub(crate) use touch::Touch;
-pub(crate) use uniq::Uniq;
-pub(crate) use url_::{UrlCommand, UrlHost, UrlPath, UrlQuery, UrlScheme};
-pub(crate) use version::Version;
-pub(crate) use where_::Where;
-pub(crate) use which_::Which;
-pub(crate) use with_env::WithEnv;
-pub(crate) use wrap::Wrap;
+pub use table::Table;
+pub use tags::Tags;
+pub use termsize::TermSize;
+pub use to::To;
+pub use to_csv::ToCsv;
+pub use to_html::ToHtml;
+pub use to_json::ToJson;
+pub use to_md::Command as ToMarkdown;
+pub use to_toml::ToToml;
+pub use to_tsv::ToTsv;
+pub use to_url::ToUrl;
+pub use to_xml::ToXml;
+pub use to_yaml::ToYaml;
+pub use touch::Touch;
+pub use uniq::Uniq;
+pub use url_::{UrlCommand, UrlHost, UrlPath, UrlQuery, UrlScheme};
+pub use version::Version;
+pub use where_::Where;
+pub use which_::Which;
+pub use with_env::WithEnv;
+pub use wrap::Wrap;
 
 #[cfg(test)]
 mod tests {
