+++ crates/nu-data/Cargo.toml
@@ -33,12 +33,14 @@ nu-errors = { version = "0.29.1", path = "../nu-errors" }
 nu-protocol = { version = "0.29.1", path = "../nu-protocol" }
 nu-source = { version = "0.29.1", path = "../nu-source" }
 nu-table = { version = "0.29.1", path = "../nu-table" }
-nu-test-support = { version = "0.29.1", path = "../nu-test-support" }
 nu-value-ext = { version = "0.29.1", path = "../nu-value-ext" }
 nu-ansi-term = { version = "0.29.1", path = "../nu-ansi-term" }
 
 [target.'cfg(unix)'.dependencies]
 users = "0.11.0"
+
+[dev-dependencies]
+nu-test-support = { version = "0.29.1", path = "../nu-test-support" }
 
 [features]
 directories = ["directories-next"]
