+++ crates/nu-engine/Cargo.toml
@@ -53,6 +53,7 @@ users = "0.11.0"
 
 [dev-dependencies]
 nu-test-support = { version = "0.29.1", path = "../nu-test-support" }
+nu-protocol = { version = "0.29.1", path = "../nu-protocol" }
 hamcrest2 = "0.3.0"
 
 [features]
