+++ crates/nu-parser/Cargo.toml
@@ -26,6 +26,8 @@ dunce = "1.0.1"
 nu-errors = { version = "0.29.1", path = "../nu-errors" }
 nu-protocol = { version = "0.29.1", path = "../nu-protocol" }
 nu-source = { version = "0.29.1", path = "../nu-source" }
+
+[dev-dependencies]
 nu-test-support = { version = "0.29.1", path = "../nu-test-support" }
 
 [features]
