+++ crates/nu-plugin/Cargo.toml
@@ -12,13 +12,6 @@ doctest = false
 [dependencies]
 nu-errors = { path = "../nu-errors", version = "0.29.1" }
 nu-protocol = { path = "../nu-protocol", version = "0.29.1" }
-nu-source = { path = "../nu-source", version = "0.29.1" }
-nu-test-support = { path = "../nu-test-support", version = "0.29.1" }
-nu-value-ext = { path = "../nu-value-ext", version = "0.29.1" }
-
-bigdecimal = { version = "0.2.0", features = ["serde"] }
-indexmap = { version = "1.6.1", features = ["serde-1"] }
-num-bigint = { version = "0.3.1", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
